\docType{data}
\name{incomplete_wisdom}
\alias{incomplete_wisdom}
\title{Phrases upon test failures.}
\format{chr [1:4] "is an obstacle to your awakening!" "requires further contemplation." "is damaging your karma." ...}
\usage{
  incomplete_wisdom
}
\description{
  Phrases upon test failures.
}
\keyword{datasets}

