\docType{data}
\name{koans}
\alias{koans}
\title{list of koans to run
context and title included to hopefully run from a single loan file (when I get around to it!)}
\format{List of 2
 $ :List of 4
 $ :List of 4}
\usage{
  koans
}
\description{
  list of koans to run context and title included to
  hopefully run from a single loan file (when I get around
  to it!)
}
\keyword{datasets}

