\name{test_all_koans}
\alias{test_all_koans}
\title{Runs checks on all answer files and koan files
This needs to return true before new koans are ready to be applied}
\usage{
  test_all_koans(koan_file = "R/koans.R",
    answers_file = "answers/koans.R")
}
\description{
  Runs checks on all answer files and koan files This needs
  to return true before new koans are ready to be applied
}

