\docType{data}
\name{koans}
\alias{koans}
\title{list of koans to run
context and title included to run from a single koan template file}
\format{List of 6
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3}
\usage{
  koans
}
\description{
  list of koans to run context and title included to run
  from a single koan template file
}
\keyword{datasets}

